doc: Relocate bindings/pydrake/doc to doc/pydrake
We want all of the website generation to live inside the doc folder.
