--- bindings/pydrake/doc/README.md
@@ -1,6 +1,0 @@
-# Python Binding Documentation
-
-To build and view documentation:
-
-    cd drake
-    bazel run //bindings/pydrake/doc:serve_sphinx
