+++ doc/README.txt
@@ -13,6 +13,7 @@ doc/
 ├── _includes/ - Jekyll include components (via Liquid). Can render Markdown.
 ├── _layouts/ - Jekyll HTML layouts. Can *not* render Markdown.
 ├── _pages/ - Jekyll collection: Add'l root-level pages. Can render Markdown.
+├── pydrake/ - Python API reference.
 ├── _release-notes/ - Jekyll collection: Index of versioned releases.
 ├── third_party/ - Third party assets.
 └── index.md - Home page.
