+++ doc/pydrake/BUILD.bazel
@@ -30,15 +30,23 @@ _DEFAULT_TEST_TAGS = [
     "block-network",
 ]
 
+drake_py_library(
+    name = "sphinx_base",
+    srcs = ["sphinx_base.py"],
+    data = ["@sphinx//:sphinx-build"],
+    tags = _DEFAULT_BINARY_TAGS,
+    deps = ["@rules_python//python/runfiles"],
+)
+
 # TODO(eric.cousineau): Add a smaller stub library stub to test this extension.
 drake_py_library(
     name = "pydrake_sphinx_extension_py",
     srcs = ["pydrake_sphinx_extension.py"],
     tags = _DEFAULT_BINARY_TAGS,
     deps = [
+        ":sphinx_base",
         "//bindings/pydrake/common:cpp_template_py",
         "//bindings/pydrake/common:deprecation_py",
-        "//doc:sphinx_base",
         "//doc:system_doxygen",
     ],
 )
@@ -62,8 +70,8 @@ drake_py_binary(
     visibility = ["//doc:__pkg__"],
     deps = [
         ":pydrake_sphinx_extension_py",
+        ":sphinx_base",
         "//bindings/pydrake",  # We must import `pydrake` to document it.
-        "//doc:sphinx_base",
     ],
 )
 
@@ -73,7 +81,7 @@ drake_py_binary(
     data = [":gen_sphinx"],
     tags = _DEFAULT_BINARY_TAGS,
     visibility = ["//doc:__pkg__"],
-    deps = ["//doc:sphinx_base"],
+    deps = [":sphinx_base"],
 )
 
 add_lint_tests_pydrake(
