+++ doc/pydrake/LICENSE.TXT
