+++ doc/pydrake/conf.py
