+++ doc/pydrake/serve_sphinx.py
@@ -4,7 +4,7 @@ Serves documentation for `pydrake`.
 
 from os.path import abspath, dirname, join
 
-from drake.doc.sphinx_base import preview_main
+from drake.doc.pydrake.sphinx_base import preview_main
 
 if __name__ == "__main__":
     gen_script = join(dirname(abspath(__file__)), "gen_sphinx")
