+++ doc/pydrake/sphinx.LICENSE.txt
